fix: fix some types not working when using package
Previously I thought it was just a problem with the `keys` function, but that was only because it was the one I used the most.

What was actually happening was that since tsc never resolves path aliases for outputted file, not even for type definitions (which is all tsc is used for in this package), when used from a package, types imported by functions in this package could not be correctly resolved.

Wrote a separate package, dts-alias, to fix the issue. It also explains the problem in depth.