refactor: changed back to old way of building types
Given the new way still required the types fix script because not all the vite plugins
are compatible, it was too complicated.

I think keeping it to simple build scripts that can be called on their on is better and also
doesn't hide what's happening (tsc/vue-tsc for generation + tsc-alias to fix the types).