refactor: improved run error handling
Since a program can write to stderr at any point and not exit with a none-zero code run now errors
if a program writes to stderr.

Technically it might not have errored, but it's easier to check in a catch
if the exit code was zero and return .data, than to change the function's return type.