feat: made pushIfNotIn slightly more performant for the average case
Usually we want to push another array, not values, and spreading is slow,
and the old way the function required spreading out arrays then internally,
it would gather them back up.

Now in the worst case we have to wrap a few values in an array and this array
isn't re-collected internally.

BREAKING CHANGE: pushIfNotIn no longer takes in non array values