feat: removed Result namespace
BREAKING CHANGE: Dropped Result namespace usage. This is to be compatible with
typescript's new `erasableSyntaxOnly` option. It also makes the package
completely tree-shakeable now.

While this is a breaking change, it should pretty much work out of the box if
you're importing the parts of the namespace individually. Just change the
`Result` import to a type import (`type Result`).

If you're doing `Result.Ok/Err`, etc, change the import to `import * as Result
...`.

Then you'll need to change any references to `Result` to `Result.Result` or the
more sensible `Result.Type` alias.